--- src/main/java/org/cqfn/astranaut/core/exceptions/BaseException.java
@@ -1,45 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2024 Ivan Kniazkov
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-package org.cqfn.astranaut.core.exceptions;
-
-/**
- * Helper class for custom exceptions in projects that use the core as a dependency.
- *
- * @since 1.0.2
- */
-public abstract class BaseException extends Exception {
-    private static final long serialVersionUID = -636126348743667929L;
-
-    /**
-     * Returns initiator name, i.e. module that thrown the exception.
-     * @return A name
-     */
-    public abstract String getInitiator();
-
-    /**
-     * Returns error message.
-     * @return A message
-     */
-    public abstract String getErrorMessage();
-}
